feat: 폴더 리팩토링 (#10)
