+++ src/domains/CompanySignup/index.ts
@@ -1,0 +1,1 @@
+export { CompanySignUpView } from "./components/CompanySingupView";
